+++ sakura_core/types/Type_Perl.cpp
@@ -409,8 +409,7 @@ const wchar_t* g_ppszKeywordsPERL[] = {
 	L"warn",
 	L"write"
 };
-int g_nKeywordsPERL = _countof(g_ppszKeywordsPERL);
-
+size_t g_nKeywordsPERL = _countof(g_ppszKeywordsPERL);
 
 // Jul. 10, 2001 JEPRO	変数を第２強調キーワードとして分離した
 // 2008/05/05 novice 重複文字列削除
@@ -517,5 +516,5 @@ const wchar_t* g_ppszKeywordsPERL2[] = {
 	L"$ENV",
 	L"$SIG"
 };
-int g_nKeywordsPERL2 = _countof(g_ppszKeywordsPERL2);
+size_t g_nKeywordsPERL2 = _countof(g_ppszKeywordsPERL2);
