+++ sakura_core/types/Type_Rich.cpp
@@ -131,5 +131,5 @@ const wchar_t* g_ppszKeywordsRTF[] = {
 	L"eml",
 	L"emr"
 };
-int g_nKeywordsRTF = _countof(g_ppszKeywordsRTF);
+size_t g_nKeywordsRTF = _countof(g_ppszKeywordsRTF);
