+++ sakura_core/types/Type_Tex.cpp
@@ -705,7 +705,7 @@ const wchar_t* g_ppszKeywordsTEX[] = {
 //			"\\}",
 //			"\\~",
 };
-int g_nKeywordsTEX = _countof(g_ppszKeywordsTEX);
+size_t g_nKeywordsTEX = _countof(g_ppszKeywordsTEX);
 
 // Jan. 19, 2001 JEPRO	TeX のキーワード2として新規追加 & 一部復活 --環境コマンドとオプション名が中心
 const wchar_t* g_ppszKeywordsTEX2[] = {
@@ -829,5 +829,5 @@ const wchar_t* g_ppszKeywordsTEX2[] = {
 //		"zh",
 //		"zw"
 };
-int g_nKeywordsTEX2 = _countof(g_ppszKeywordsTEX2);
+size_t g_nKeywordsTEX2 = _countof(g_ppszKeywordsTEX2);
